load/requirements.txt
aenum==2.2.6
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
frozenlist==1.2.0
greenlet==1.1.2
gremlinpython==3.5.2
idna==3.3
install==1.3.5
isodate==0.6.1
loguru==0.5.3
minio==7.1.2
multidict==5.2.0
nest-asyncio==1.5.4
numpy==1.21.5
pandas==1.3.5
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
pyyaml==5.4.1
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.8
yarl==1.7.2
